feat:file check

+++ ExtremeSkins.Converter.Core/ExtremeHatConverter.cs
@@ -42,7 +42,7 @@ public sealed class ExtremeVisorConverter : ICosmicConverter
 
     private static bool TryCopyFile(string outputFile, string targetFile)
     {
-        if (!string.IsNullOrEmpty(targetFile))
+        if (!string.IsNullOrEmpty(targetFile) && File.Exists(targetFile))
         {
             File.Copy(targetFile, outputFile);
             return true;
